+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 - feat: make testsuite and run testsuite
 - feat: testcase/testsuite config support getting variables by function
 - feat: har2case with request cookies
+- feat: log request/response headers and body with indent
 
 **Fixed**